Create Test case for service testing
This commit will allow us to reuse the mock object in different test
files inside the services package.
